Moved gauge series to subfolder
